+++ BlackCat/Magazine/MagazineMock/MagazineModel.dummy.swift
